requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
openai==1.3.7
python-multipart==0.0.6
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1
schedule==1.2.0
